feat: Alteração da renda de int para double.

+++ src/main/java/desafioTec/digix/criteria/CriterioRenda901Ate1500.java
@@ -6,7 +6,7 @@ public class CriterioRenda901Ate1500 implements CriterioPontuacao{
 
     @Override
     public int calcularPontucao(Familia familia) {
-        int rendaFamilia = familia.getRendaTotal();
+        double rendaFamilia = familia.getRendaTotal();
         return rendaFamilia > 900 & rendaFamilia <= 1500 ? 3 : 0;
     }
 }
